feat: Change default model dtype to bfloat16

+++ src/data/config.py
@@ -57,7 +57,7 @@ class ModelConfig:
     sigma_min: float = 0.002
     sigma_max: float = 80.0
     rho: float = 7.0
-    dtype: str = "float32"  # Model precision: float32, float16, or bfloat16
+    dtype: str = "bfloat16"  # Model precision: float32, float16, or bfloat16
     fallback_dtype: str = "float32"  # Fallback precision when main dtype not supported
     
 @dataclass
